refactor: extract BinContextMenuWrapper to shared component
Move BinContextMenuWrapper from being duplicated in both App.tsx and
MobileLayout.tsx to a shared component in components/mobile/.

The component handles routing between single-bin and multi-bin context
menus based on current selection state. Having it in one place ensures
consistent behavior and reduces the risk of fixes being applied to
only one location.
